backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
neo4j==5.16.0
redis==5.0.1

# Vector Database / Embeddings
pgvector==0.2.4
sentence-transformers==2.2.2
huggingface-hub==0.20.3

# LLM Integration
openai==1.10.0
langchain==0.1.4
langchain-community==0.0.16

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.2
beautifulsoup4==4.12.3
 unstructured[all-docs]==0.12.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.26.0
aiofiles==23.2.1
celery==5.3.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
